Refactored server for future testing
